# src/main/resources/openapi/payment-api.yaml
openapi: 3.0.3
info:
  title: Payment Service API
  version: 1.0.0
  description: API for managing payments

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /payments:
    get:
      summary: List payments with optional filters and pagination
      operationId: listPayments
      tags:
        - payments
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - PENDING
              - APPROVED
              - FAILED
              - CANCELED
          description: Filter by payment status
          example: PENDING
        - name: userId
          in: query
          required: false
          schema:
            type: string
          description: Filter by user identifier
          example: user123
        - name: orderId
          in: query
          required: false
          schema:
            type: string
          description: Filter by order identifier
          example: order456
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0
          description: Page number (0-indexed)
          example: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
          description: Page size (1-100)
          example: 20
      responses:
        '200':
          description: Payments retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentListResponse'
        '400':
          description: Bad request - invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new payment
      operationId: createPayment
      tags:
        - payments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: Payment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /payments/{id}:
    get:
      summary: Retrieve a payment by ID
      operationId: getPaymentById
      tags:
        - payments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Payment identifier
          example: 1
      responses:
        '200':
          description: Payment retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Bad request - invalid ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 409
                message: Invalid status transition
                details: Cannot transition from terminal state APPROVED to PENDING. Payments in APPROVED status cannot be modified.
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /payments/{id}/status:
    patch:
      summary: Update payment status
      operationId: updatePaymentStatus
      tags:
        - payments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Payment identifier
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum:
                    - APPROVED
                    - FAILED
                    - CANCELED
                  description: New payment status
                  example: APPROVED
      responses:
        '200':
          description: Payment status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Bad request - invalid status value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - invalid status transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 409
                message: Invalid status transition
                details: Cannot transition from terminal state APPROVED to PENDING. Payments in APPROVED status cannot be modified.
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CreatePaymentRequest:
      type: object
      required:
        - amount
        - currency
        - method
        - userId
      properties:
        amount:
          type: number
          format: double
          minimum: 0.01
          description: Payment amount (must be greater than 0)
          example: 99.99
        currency:
          type: string
          minLength: 1
          description: Currency code (ISO 4217 format)
          example: USD
        method:
          type: string
          enum:
            - CREDIT_CARD
            - DEBIT_CARD
            - PAYPAL
            - BANK_TRANSFER
          description: Payment method
          example: CREDIT_CARD
        userId:
          type: string
          minLength: 1
          description: User identifier
          example: user123
        orderId:
          type: string
          nullable: true
          description: Optional order identifier
          example: order456

    PaymentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique payment identifier
          example: 1
        amount:
          type: number
          format: double
          description: Payment amount
          example: 99.99
        currency:
          type: string
          description: Currency code
          example: USD
        method:
          type: string
          description: Payment method
          example: CREDIT_CARD
        status:
          type: string
          enum:
            - PENDING
            - APPROVED
            - FAILED
            - CANCELED
          description: Payment status
          example: PENDING
        userId:
          type: string
          description: User identifier
          example: user123
        orderId:
          type: string
          nullable: true
          description: Order identifier
          example: order456
        createdAt:
          type: string
          format: date-time
          description: Payment creation timestamp
          example: 2024-01-15T10:30:00Z
        updatedAt:
          type: string
          format: date-time
          description: Payment last update timestamp
          example: 2024-01-15T10:30:00Z

    PaymentListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PaymentResponse'
          description: List of payments
        totalElements:
          type: integer
          format: int64
          description: Total number of payments matching the filters
          example: 100
        totalPages:
          type: integer
          format: int32
          description: Total number of pages
          example: 5
        page:
          type: integer
          format: int32
          description: Current page number (0-indexed)
          example: 0
        size:
          type: integer
          format: int32
          description: Page size
          example: 20

    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Error message
          example: Validation failed
        details:
          type: string
          description: Additional error details
          example: Amount must be greater than 0